docker/memory/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
memu-py[postgres]>=1.4.0
pydantic-settings>=2.0.0
